requirements.txt
annotated-types==0.7.0
black==25.1.0
certifi==2025.4.26
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
distlib==0.3.9
exceptiongroup==1.3.0
filelock==3.18.0
identify==2.6.12
idna==3.10
iniconfig==2.1.0
isort==6.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
shellingham==1.5.4
six==1.17.0
tomli==2.2.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.31.2
